Updated decryption
So there's an output now and it has tonal range & lacks NaN values. Problem is I'm pretty sure we're gonna struggle with this. OG vs return_OG is pretty wild & not good.

Maybe with lots of debugging/confirming no major errors this can be fixed. But it also might just not work tbh