Issue==
This decryption is running into a big issue when it comes to key_to_img2. It seems that the key is being recognized as shorter than the image input (despite this not being the case), and is then being trimmed down. That means the arrays are not the same and further operations can't run. WIll see what I can do with this later